deleted appold and moved the spotify code into the new login routes
