+++ Server/package.json
@@ -5,7 +5,8 @@
     "mongodb": "^4.1.3",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.14",
-    "pg": "^8.7.1"
+    "pg": "^8.7.1",
+    "spotify-web-api-node": "^5.0.2"
   },
   "name": "Server",
   "version": "1.0.0",
